Make volatile coroutines work a little better
If they're registered with an endpoint, at least
minimize the chance that two endpoints clash with
their IDs (and attempt to resume coroutines with
state that doesn't match). If an coroutine instance
can't be found, return IncompatibleState so the
call is restarted rather than discarded (as is the
case with NotFound).